TVIST1-297: Renamed agenda types to be more telling
