+++ src/Form/AgendaNewType.php
@@ -4,55 +4,38 @@ namespace App\Form;
 
 use App\Entity\Agenda;
 use App\Entity\Board;
-use App\Service\AgendaStatus;
+use Doctrine\ORM\EntityRepository;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\DateType;
-use Symfony\Component\Form\Extension\Core\Type\SubmitType;
-use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\Extension\Core\Type\TimeType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\Contracts\Translation\TranslatorInterface;
 
-class AgendaType extends AbstractType
+class AgendaNewType extends AbstractType
 {
-    /**
-     * @var TranslatorInterface
-     */
-    private $translator;
-
-    public function __construct(TranslatorInterface $translator)
-    {
-        $this->translator = $translator;
-    }
-
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             'data_class' => Agenda::class,
+            'municipality' => null,
         ]);
     }
 
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
+        $municipality = $options['municipality'];
+
         $builder
             ->add('board', EntityType::class, [
                 'class' => Board::class,
+                'query_builder' => function (EntityRepository $er) use ($municipality) {
+                    return $er->createQueryBuilder('b')
+                        ->where('b.municipality = :municipality')
+                        ->setParameter('municipality', $municipality->getId()->toBinary())
+                        ->orderBy('b.name', 'ASC');
+                },
                 'choice_label' => 'name',
-                'disabled' => true,
-            ])
-            ->add('status', ChoiceType::class, [
-                'choices' => [
-                    $this->translator->trans('Open', [], 'agenda') => AgendaStatus::OPEN,
-                    $this->translator->trans('Full', [], 'agenda') => AgendaStatus::FULL,
-                    $this->translator->trans('Ready', [], 'agenda') => AgendaStatus::READY,
-                    $this->translator->trans('Finished', [], 'agenda') => AgendaStatus::FINISHED,
-                ],
-            ])
-            ->add('remarks', TextareaType::class, [
-                'required' => false,
             ])
             ->add('date', DateType::class, [
                 'widget' => 'single_text',
@@ -68,8 +51,6 @@ class AgendaType extends AbstractType
                 'widget' => 'single_text',
                 'input_format' => 'H:i',
             ])
-            ->add('submit', SubmitType::class, [
-                'label' => $this->translator->trans('Save changes', [], 'agenda'),
-            ]);
+        ;
     }
 }
